requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
branca==0.7.1
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dm-tree==0.1.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
eval_type_backport==0.2.2
fastapi==0.110.0
filelock==3.13.1
flatbuffers==23.5.26
folium==0.16.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.3
humanfriendly==10.0
hume==0.7.7
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
Jinja2==3.1.3
jiter==0.8.0
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.0.5
kiwisolver==1.4.5
kubernetes==29.0.0
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.9.2
llama-index==0.9.44
llama-index-agent-openai==0.1.1
llama-index-cli==0.1.5
llama-index-core==0.10.68.post1
llama-index-embeddings-openai==0.1.1
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.1
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.1
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.4
llama-parse==0.3.4
llamaindex-py-client==0.1.13
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.56.1
openpyxl==3.1.5
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
posthog==3.4.2
preshed==3.0.9
protobuf==4.25.3
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.2
pypdf==4.0.2
PyPDF2==3.0.1
pyphen==0.14.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.3
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
simpleaudio==1.0.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.27
srsly==2.4.8
srt==3.5.3
starlette==0.36.3
streamlit==1.31.1
streamlit-audiorec==0.1.3
streamlit-folium==0.18.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
textblob==0.18.0.post0
textstat==0.7.3
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.67.1
transformers==4.39.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
vosk==0.3.44
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
